fix issue #379 (https://github.com/Pointcept/Pointcept/issues/379#issue-2731106225) (#399)

+++ pointcept/utils/scheduler.py
@@ -24,7 +24,7 @@ class MultiStepLR(lr_scheduler.MultiStepLR):
     ):
         super().__init__(
             optimizer=optimizer,
-            milestones=[rate * total_steps for rate in milestones],
+            milestones=[int(rate * total_steps) for rate in milestones],
             gamma=gamma,
             last_epoch=last_epoch,
             verbose=verbose,
